pyproject.toml
[project]
name = "lockeroo-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Manuel Rottschaefer",email = "manuel@lockeroo.de"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.7.0)",
    "argon2-cffi (==23.1.0)",
    "argon2-cffi-bindings (==21.2.0)",
    "bcrypt (==4.2.1)",
    "beanie (==1.26.0)",
    "blinker (==1.9.0)",
    "brotli (==1.1.0)",
    "certifi (==2024.8.30)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.0)",
    "click (==8.1.8)",
    "configargparse (==1.7)",
    "cryptography (==44.0.0)",
    "dnspython (==2.7.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.6)",
    "fastapi-mqtt (==2.2.0)",
    "fastapi-users (==14.0.0)",
    "fastapi-users-db-beanie (==3.0.0)",
    "fief-client (==0.20.0)",
    "flask (==3.1.0)",
    "flask-cors (==5.0.0)",
    "flask-login (==0.6.3)",
    "gevent (==24.11.1)",
    "geventhttpclient (==2.3.3)",
    "gmqtt (==0.7.0)",
    "greenlet (==3.1.1)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httpx (==0.27.2)",
    "idna (==3.10)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.4)",
    "jwcrypto (==1.5.6)",
    "lazy-model (==0.2.0)",
    "loguru (==0.7.2)",
    "locust (==2.32.4)",
    "makefun (==1.15.6)",
    "markupsafe (==3.0.2)",
    "motor (==3.6.0)",
    "msgpack (==1.1.0)",
    "paho-mqtt (==2.1.0)",
    "pip (==24.3.1)",
    "psutil (==6.1.0)",
    "pwdlib (==0.2.1)",
    "pycparser (==2.22)",
    "pydantic (==2.10.4)",
    "pydantic-core (==2.27.2)",
    "pydeps (==2.0.1)",
    "pygyat (==1.0.6b0)",
    "pyjwt (==2.9.0)",
    "pymongo (==4.9.2)",
    "python-dotenv (==1.0.1)",
    "python-multipart (==0.0.17)",
    "pyyaml (==6.0.2)",
    "pyzmq (==26.2.0)",
    "requests (==2.32.3)",
    "setuptools (==75.6.0)",
    "sniffio (==1.3.1)",
    "starlette (==0.41.3)",
    "stdlib-list (==0.11.0)",
    "toml (==0.10.2)",
    "typing-extensions (==4.12.2)",
    "urllib3 (==2.2.3)",
    "uvicorn (==0.32.1)",
    "websocket (==0.2.1)",
    "websockets (==14.1)",
    "werkzeug (==3.1.3)",
    "zope-event (==5.0)",
    "zope-interface (==7.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
